Доработать архитектуру
