+++ internal/server/model/metric.go
@@ -1,0 +1,7 @@
+package model
+
+type Metric struct {
+	MetType  string
+	MetName  string
+	MetValue any
+}
