+++ internal/server/server/const.go
@@ -1,0 +1,9 @@
+package server
+
+const (
+	GaugeType    = "gauge"
+	CounterType  = "counter"
+	Addr         = ":8080"
+	ContentType  = "text/plain"
+	UpdateMethod = "update"
+)
