+++ internal/server/storage/store.go
@@ -1,0 +1,7 @@
+// package storage
+
+// import "github.com/golovanevvs/metalecoll/internal/server/model"
+
+// type Repository interface {
+// 	metricSave(m model.Metric)
+// }
